chore: refactor permission for sandBox
