+++ src/App.vue
@@ -14,7 +14,7 @@
 export default {
   name: 'App',
   mounted() {
-    this.$dialog({ title: '注意', content: '版本更新后请清除localStorage数据！' })
+    // this.$dialog({ title: '注意', content: '版本更新后请清除localStorage数据！' })
   }
 };
 </script>
