+++ src/ppsUI/theme/contextMenu.scss
@@ -8,7 +8,7 @@
     border-radius: var(--radius);
     box-shadow: 0 2px 8px 0 rgba(0, 0, 0, 0.145);
     background: #ffffff;
-    padding: 10px;
+    padding: 6px;
     z-index: 100;
     color: initial;
     .pps-context-menu_prepend {
@@ -37,6 +37,7 @@
     .pps-context-menu_inner {
       .pps-context-menu-item {
         padding: 5px 10px;
+        border-radius: inherit;
         cursor: pointer;
         &:hover {
           background: #f5f5f5;
